moves tests to depreacted directory
