+++ src/__tests__/deprecated/FormGroup.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import FormGroup from '../deprecated/FormGroup'
-import {behavesAsComponent, checkExports} from '../utils/testing'
+import FormGroup from '../../deprecated/FormGroup'
+import {behavesAsComponent, checkExports} from '../../utils/testing'
 import {render as HTMLRender, cleanup} from '@testing-library/react'
 import {axe, toHaveNoViolations} from 'jest-axe'
 import 'babel-polyfill'
