+++ src/__tests__/deprecated/InputField.test.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import {render, cleanup} from '@testing-library/react'
 import {axe, toHaveNoViolations} from 'jest-axe'
 import 'babel-polyfill'
-import {Autocomplete, SSRProvider, TextInput, TextInputWithTokens} from '..'
-import InputField from '../deprecated/InputField'
+import {Autocomplete, SSRProvider, TextInput, TextInputWithTokens} from '../../'
+import InputField from '../../deprecated/InputField'
 expect.extend(toHaveNoViolations)
 
 const TEXTINPUTFIELD_LABEL_TEXT = 'Name'
